+++ examples/example/spec/test_pages/page1.html
@@ -1,6 +1,5 @@
 <html>
-<body>
-  <div><div><div><div id="element" class ="1 2 3 4 5">Text</div></div></div></div>
-
+  <body>
+    <div><div><div><div id="element" class ="1 2 3 4 5">Text</div></div></div></div>
   </body>
-  </html>
+</html>
