+++ lib/locatine.rb
@@ -1,5 +1,15 @@
 # frozen_string_literal: true
 
+require 'locatine/daemon_helpers/methods'
+require 'locatine/results_helpers/common'
+require 'locatine/results_helpers/find_by_magic'
+require 'locatine/results_helpers/guess'
+require 'locatine/results_helpers/info_generator'
+require 'locatine/results_helpers/xpath_generator'
+
+require 'locatine/logger'
 require 'locatine/version'
 require 'locatine/daemon'
 require 'locatine/session'
+require 'locatine/element'
+require 'locatine/results'
